+++ packages/postgres-database/tests/test_delete_projects_and_users.py
@@ -9,7 +9,7 @@ import pytest
 import sqlalchemy as sa
 from aiopg.sa.result import ResultProxy, RowProxy
 from psycopg2.errors import ForeignKeyViolation  # pylint: disable=no-name-in-module
-from pytest_simcore.helpers.fakers_rawdata_for_db import random_project, random_user
+from pytest_simcore.helpers.rawdata_fakers import random_project, random_user
 from simcore_postgres_database.models.base import metadata
 from simcore_postgres_database.webserver_models import projects, users
