+++ packages/postgres-database/tests/test_groups.py
@@ -8,7 +8,7 @@ import aiopg.sa.exc
 import pytest
 from aiopg.sa.result import ResultProxy, RowProxy
 from psycopg2.errors import ForeignKeyViolation, RaiseException, UniqueViolation
-from pytest_simcore.helpers.fakers_rawdata_for_db import random_group, random_user
+from pytest_simcore.helpers.rawdata_fakers import random_group, random_user
 from simcore_postgres_database.models.base import metadata
 from simcore_postgres_database.webserver_models import (
     GroupType,
